I understood my mistake when I was laying in my bed, and fixed the problem

+++ src/com/taraschuiko/labs/Lab1.java
@@ -5,22 +5,16 @@ import java.util.Scanner;
 public class Lab1 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-
         System.out.print("a: ");
         double a = scanner.nextDouble();
-
         System.out.print("b: ");
         double b = scanner.nextDouble();
-
         System.out.print("c: ");
         double c = scanner.nextDouble();
-
         System.out.print("X start: ");
         double xStart = scanner.nextDouble();
-
         System.out.print("X end: ");
         double xEnd = scanner.nextDouble();
-
         System.out.print("dX: ");
         double dx = scanner.nextDouble();
 
@@ -28,7 +22,7 @@ public class Lab1 {
             if (!((int) a != 0 || (int) b != 0 || (int) c != 0)) {
                 f(a, b, c, x);
             } else {
-                f(a, b, c, x);
+                f((int) a, (int) b, (int) c, (int) x);
             }
         }
     }
